✨ feat: Refatoração da configuração do Axios e atualização de importaes para o novo serviço
