+++ src/pages/ResetPassword.tsx
@@ -2,7 +2,7 @@ import React, { useState } from 'react';
 import { useSearchParams, useNavigate, Link } from 'react-router-dom';
 import Typography from '../components/ui/Typography';
 import SubmitButton from '../components/ui/SubmitButton';
-import api from '../config/axiosConfig';
+import api from '../services/api';
 import { toast } from 'react-toastify';
 
 const ResetPassword: React.FC = () => {
