+++ src/services/api.ts
@@ -1,7 +1,17 @@
 import axios from 'axios';
 
+// Configuração global do Axios
 const api = axios.create({
-  baseURL: 'http://localhost:3000',
+  baseURL: import.meta.env.VITE_API_URL,
+});
+
+// Adicionar headers globais (exemplo: autenticação)
+api.interceptors.request.use((config) => {
+  const token = localStorage.getItem('token');
+  if (token) {
+    config.headers.Authorization = `Bearer ${token}`;
+  }
+  return config;
 });
 
 export default api;
